Can conditionally add fields to support cross-collection collections
