+++ MARC_set_wrangler.rb
@@ -458,53 +458,6 @@ def clean_id(rec, idfields, spec)
   return rec
 end
 
-class MarcEdit
-  def initialize(rec)
-    @rec = rec
-  end
-
-  def add_field(fspec)
-    fs = fspec.dup
-    field = MARC::DataField.new(fs['tag'], fs['i1'], fs['i2'])
-    fs['subfields'].each do |sf|
-      field.append(MARC::Subfield.new(sf['delimiter'], sf['value']))
-    end
-    @rec.append(field)
-  end
-
-  def add_field_with_parameter(fspec, replaces)
-    fspec.each { |fs|
-      sfval = ''
-      f = MARC::DataField.new(fs['tag'], fs['i1'], fs['i2'])
-      fs['subfields'].each { |sfs|
-        sfval = sfs['value'].dup
-        if replaces.size > 0
-          replaces.each { |findrep|
-            findrep.each_pair { |fnd, rep|
-              sfval.gsub!(fnd, rep)
-            }
-          }
-        end
-        sf = MARC::Subfield.new(sfs['delimiter'], sfval)
-        f.append(sf)
-      }
-      @rec.append(f)
-    }
-  end
-
-  def sort_fields
-    old_ldr = @rec.leader
-    field_hash = @rec.group_by { |field| field.tag }
-    newrec = MARC::Record.new()
-    field_hash.keys.sort!.each do |tag|
-      field_hash[tag].each { |f| newrec.append(f) }
-    end
-    newrec.leader = old_ldr
-    @rec = newrec
-    @rec
-  end
-end
-
 def check_for_multiple_overlays(sets)
   errs = []
   sets.each { |set|
@@ -807,6 +760,12 @@ in_info.each_value do |ri|
 
   if thisconfig['add MARC field spec']
     thisconfig['add MARC field spec'].each { |field_spec| reced.add_field(field_spec) }
+  end
+
+  if thisconfig['add conditional MARC field with parameters spec']
+    thisconfig['add conditional MARC field with parameters spec'].each do |field_spec|
+      reced.add_conditional_field_with_parameters(field_spec)
+    end
   end
 
   if thisconfig['write warnings to recs']
