+++ config.yaml
@@ -192,7 +192,7 @@ workflows:
       - tag: '019'
       - tag: '035'
       - tag: '049'
-      - tag: '072'    
+      - tag: '072'
       - tag: '082'
       - tag: '506'
       - tag: '590'
@@ -368,6 +368,86 @@ workflows:
 collections:
   generic collection (no special collection-level processing):
     placeholder option: true
+  WCM - CombinedMonthly (cross-collection):
+    id affix value: 'WCMcombined'
+    add MARC field spec:
+      - tag: '773'
+        i1: '1'
+        i2: ' '
+        subfields:
+          - delimiter: 't'
+            value: 'OCLC WorldShare Collection Manager managed collection. WCMcombined'
+      - tag: '506'
+        i1: '1'
+        i2: ' '
+        subfields:
+          - delimiter: 'a'
+            value: 'Access limited to UNC Chapel Hill-authenticated users.'
+          - delimiter: 'f'
+            value: 'Unlimited simultaneous users'
+    add conditional MARC field with parameters spec:
+      - find:
+          tag: '773'
+        add:
+          - tag: '590'
+            i1: '0'
+            i2: ' '
+            subfields:
+              - delimiter: 'a'
+                value: 'Content provider: provider_param.'
+        mappings:
+          - value: 'ACS symposium series.*(online collection)'
+            provider_param: 'American Chemical Society'
+          - value: 'AGU ebooks.*(online collection)'
+            provider_param: 'American Geophysical Union via Wiley Online Library'
+          - value: 'AIP Conference Proceedings.*(online collection)'
+            provider_param: 'American Institute of Physics'
+          - value: 'American Mathematical Society.*(online collection)'
+            provider_param: 'American Mathematical Society (AMS)'
+          - value: 'Business Expert Press.*(online collection)'
+            provider_param: 'IGPublish'
+          - value: 'Cambridge histories.*(online collection)'
+            provider_param: 'Cambridge University Press'
+          - value: 'Cambridge University Press.*(online collection)'
+            provider_param: 'Cambridge University Press'
+          - value: 'CIAO.*(online collection)'
+            provider_param: 'CIAO'
+          - value: 'Columbia University Press.*(online collection)'
+            provider_param: 'De Gruyter'
+          - value: 'Cornell University Press.*(online collection)'
+            provider_param: 'De Gruyter'
+          - value: 'Credo Academic.*(online collection)'
+            provider_param: 'Credo Reference'
+          - value: 'de Gruyter.*(online collection)'
+            provider_param: 'De Gruyter'
+          - value: 'European Mathematical Society.*(online collection)'
+            provider_param: 'European Mathematical Society'
+          - value: 'Film Platform.*(online collection)'
+            provider_param: 'Film Platform'
+          - value: 'GeoScienceWorld.*(online collection)'
+            provider_param: 'GeoScienceWorld'
+          - value: 'Harvard University Press.*(online collection)'
+            provider_param: 'De Gruyter'
+          - value: 'Medici.tv.*(online collection)'
+            provider_param: 'Medici.tv'
+          - value: 'Numerique premium.*(online collection)'
+            provider_param: 'Numerique Premium'
+          - value: 'Princeton University Press.*(online collection)'
+            provider_param: 'De Gruyter'
+          - value: 'Psychotherapy.net*(online collection)'
+            provider_param: 'Psychotheraphy.net'
+          - value: 'Royal Society of Chemistry.*(online collection)'
+            provider_param: 'Royal Society of Chemistry (Great Britain)'
+          - value: 'Rutgers University Press.*(online collection)'
+            provider_param: 'De Gruyter'
+          - value: 'SPIE ebooks.*(online collection)'
+            provider_param: 'SPIE'
+          - value: 'University of Hawaii.*(online collection)'
+            provider_param: 'De Gruyter'
+          - value: 'University of Pennsylvania.*(online collection)'
+            provider_param: 'De Gruyter'
+          - value: 'University of Toronto.*(online collection)'
+            provider_param: 'De Gruyter'
   WCM - ACS:
     id affix value: 'ACS'
     add MARC field spec:
@@ -1103,7 +1183,7 @@ collections:
           - delimiter: 'f'
             value: 'Unlimited simultaneous users'
       - tag: '590'
-        i1: ' '
+        i1: '0'
         i2: ' '
         subfields:
           - delimiter: 'a'
